r1lite_ros1/requirements.txt
actionlib==1.14.3
angles==1.9.14
apturl==0.5.2
asttokens==3.0.0
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
backcall==0.2.0
base_local_planner==1.17.3
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
bondpy==1.8.7
Brlapi==0.7.0
camera-calibration==1.17.0
camera-calibration-parsers==1.12.1
catkin==0.8.12
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
catkin-tools==0.9.4
cbor==1.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cryptography==2.8
cupshelpers==1.0
cv-bridge==1.16.2
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
devscripts===2.20.2ubuntu2
diagnostic-analysis==1.12.1
diagnostic-common-diagnostics==1.12.1
diagnostic-updater==1.12.1
distro==1.4.0
distro-info==0.23+ubuntu1.1
dnspython==1.16.0
docstring_parser==0.16
docutils==0.16
duplicity==0.8.12.0
dynamic-reconfigure==1.7.6
dynamixel_sdk==3.8.0
empy==3.3.2
entrypoints==0.3
eval_type_backport==0.2.2
evdev==1.9.2
executing==2.2.0
fasteners==0.14.1
filelock==3.0.12
fsspec==2024.6.1
future==0.18.2
gazebo_plugins==2.9.3
gazebo_ros==2.9.3
gencpp==0.7.2
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.1
gennodejs==2.0.2
genpy==0.6.18
gpg==1.13.1
gssapi==1.6.1
h5py==3.11.0
html5lib==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
image-geometry==1.16.2
importlib_metadata==8.5.0
incremental==16.10.1
interactive-markers==1.12.2
ipaclient==4.8.6
ipalib==4.8.6
ipaplatform==4.8.6
ipapython==4.8.6
ipdb==0.13.13
ipython==8.12.3
jedi==0.19.2
Jinja2==2.10.1
joint-state-publisher==1.15.2
joint-state-publisher-gui==1.15.2
kaptan==0.5.10
kdl-py==1.2.0
# Editable install with no version control (kdl_parser_py==3.0.0)
-e /home/yingbo/r1lite/lerobot_replay/kdl_parser_py-rolling
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
laser-geometry==1.6.8
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libtmux==0.8.2
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
Mako==1.1.0
markdown-it-py==3.0.0
MarkupSafe==1.1.0
matplotlib==3.1.2
matplotlib-inline==0.1.7
mdurl==0.1.2
message-filters==1.17.4
monotonic==1.5
more-itertools==4.2.0
moveit-commander==1.1.16
moveit-core==1.1.16
moveit-ros-planning-interface==1.1.16
moveit-ros-visualization==1.1.16
mpi4py==3.0.3
mpmath==1.3.0
netaddr==0.7.19
netifaces==0.10.4
networkx==3.0
nose==1.3.7
numpy==1.24.4
oauthlib==3.1.0
olefile==0.46
osrf-pycommon==2.0.2
pandas==2.0.3
paramiko==2.6.0
parso==0.8.5
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
ply==3.11
prompt_toolkit==3.0.52
protobuf==3.6.1
psutil==5.5.1
pure_eval==0.2.3
py-ubjson==0.14.0
pyarrow==17.0.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pyassimp==4.1.4
pycairo==1.16.2
pycparser==2.19
pycryptodomex==3.6.1
pycups==1.9.73
pydot==1.4.1
Pygments==2.19.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyRFC3339==1.1
pyserial==3.5
python-apt==2.0.1+ubuntu0.20.4.1
python-augeas==0.5.0
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-gnupg==0.4.5
python-ldap==3.2.0
python-magic==0.4.16
python-pcl==0.3
python-qt-binding==0.4.6
python-snappy==0.5.3
python-yubico==1.3.3
PyTrie==0.2
pytz==2025.2
pyusb==1.0.2
pyxdg==0.26
PyYAML==5.3.1
qrcode==6.1
qt-dotgraph==0.4.5
qt-gui==0.4.5
qt-gui-cpp==0.4.5
qt-gui-py-common==0.4.5
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
resource-retriever==1.12.10
rich==14.0.0
roman==2.0.0
ros_numpy==0.0.5
rosbag==1.17.4
rosboost-cfg==1.15.10
rosclean==1.15.10
roscreate==1.15.10
rosdep==0.26.0
rosdep-modules==0.26.0
rosdepc==1.1.0
rosdistro-modules==1.0.1
rosgraph==1.17.4
roslaunch==1.17.4
roslib==1.15.10
roslint==0.12.0
roslz4==1.17.4
rosmake==1.15.10
rosmaster==1.17.4
rosmsg==1.17.4
rosnode==1.17.4
rosparam==1.17.4
rospkg==1.6.0
rospkg-modules==1.6.0
rospy==1.17.4
rosservice==1.17.4
rostest==1.17.4
rostopic==1.17.4
rosunit==1.15.10
roswtf==1.17.4
rqt-action==0.4.11
rqt-bag==0.5.3
rqt-bag-plugins==0.5.3
rqt-console==0.4.14
rqt-dep==0.4.14
rqt-graph==0.4.16
rqt-gui==0.5.5
rqt-gui-py==0.5.5
rqt-image-view==0.4.19
rqt-launch==0.4.10
rqt-logger-level==0.4.13
rqt-moveit==0.5.13
rqt-msg==0.4.12
rqt-nav-view==0.5.8
rqt-plot==0.4.16
rqt-pose-view==0.5.13
rqt-publisher==0.4.12
rqt-py-common==0.5.5
rqt-py-console==0.4.12
rqt-reconfigure==0.5.7
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-robot-steering==0.5.14
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.2
rqt-service-caller==0.4.12
rqt-shell==0.4.13
rqt-srv==0.4.11
rqt-tf-tree==0.6.5
rqt-top==0.4.11
rqt-topic==0.4.15
rqt-web==0.4.11
rviz==1.14.26
scipy==1.10.1
SecretStorage==2.3.1
sensor-msgs==1.13.2
service-identity==18.1.0
shtab==1.7.2
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smach==2.5.3
smach-ros==2.5.3
smclib==1.8.7
soupsieve==1.9.5
srdfdom==0.6.4
ssh-import-id==5.10
stack-data==0.6.3
sympy==1.13.3
systemd-python==234
tf==1.13.4
tf-conversions==1.13.4
tf2-geometry-msgs==0.7.10
tf2-kdl==0.7.10
tf2-py==0.7.10
tf2-ros==0.7.10
tmuxp==1.5.4
tomli==2.2.1
topic-tools==1.17.4
torch==2.4.1+cpu
torchvision==0.19.1+cpu
trac_ik_python==1.6.7
traitlets==5.14.3
transforms3d==0.4.2
Twisted==18.9.0
txaio==2.10.0
typeguard==4.4.0
typing_extensions==4.13.2
tyro==0.9.26
tzdata==2025.2
u-msgpack-python==2.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
unidiff==0.5.5
urdf-parser-py==0.0.4
urdfdom-py==0.4.6
urllib3==1.25.8
usb-creator==0.3.7
vcstool==0.3.0
wadllib==1.3.3
wcwidth==0.2.13
webencodings==0.5.1
wsaccel==0.6.2
xacro==1.14.20
xkit==0.0.0
zipp==3.20.2
zope.interface==4.7.1
